# myApp/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Calculator Exercise",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Calculator",
            "description": "API for Calculus in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/calculator/add": {
            "get": {
                "tags": [
                    "Addition Service"
                ],
                "summary": "Add 2 numbers",
                "parameters": [
                    {
                        "name": "num1",
                        "in": "query",
                        "description": "the First Number"
                    },
                    {
                        "name": "num2",
                        "in": "query",
                        "description": "the Second number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This service allows you to add two numbers together"
                    }
                }
            }
        },
        "/calculator/subtract": {
            "get": {
                "tags": [
                    "Addition Service"
                ],
                "summary": "Subtract number",
                "parameters": [
                    {
                        "name": "num1",
                        "in": "query",
                        "description": "the First Number"
                    },
                    {
                        "name": "num2",
                        "in": "query",
                        "description": "the Second number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This service allows you to subtract the second number from the first number"
                    }
                }
            }
        },
        "/calculator/multiply": {
            "get": {
                "tags": [
                    "Addition Service"
                ],
                "summary": "Multiply 2 numbers",
                "parameters": [
                    {
                        "name": "num1",
                        "in": "query",
                        "description": "the First Number"
                    },
                    {
                        "name": "num2",
                        "in": "query",
                        "description": "the Second number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This service allows you to multiply two numbers"
                    }
                }
            }
        },
        "/calculator/divide": {
            "get": {
                "tags": [
                    "Addition Service"
                ],
                "summary": "Divide by number",
                "parameters": [
                    {
                        "name": "num1",
                        "in": "query",
                        "description": "the First Number"
                    },
                    {
                        "name": "num2",
                        "in": "query",
                        "description": "the Second number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This service allows you to divide the first number by the second number"
                    }
                }
            }
        }
    }
}